\name{RcmdrPlugin.EBM-internal}
\title{Internal RcmdrPlugin.EBM objects}
\alias{fncGetConfIntText}
\alias{fncNbPercents}
\alias{fncEBMCrossTab}
\alias{fncEBMCrossTabWin}
\alias{fncEBMTherapy}
\alias{fncEBMPrognosis}
\alias{fncEBMDiagnosis}
\alias{enterTableEBMCrossTab}
\alias{fncEBMPostTestWin}
\description{Internal RcmdrPlugin.EBM objects.}
\details{These are not to be called by the user.}
\keyword{internal}
