\name{RcmdrPlugin.EBM-package}
\alias{RcmdrPlugin.EBM-package}
\alias{RcmdrPlugin.EBM}
\docType{package}
\title{
Evidence Based Medicine Rcmdr Plug-In
}
\description{
This package provides an Rcmdr plug-in for Evidence Based Medicine calculations.
}
\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.EBM\cr
Type: \tab Package\cr
Version: \tab 1.0-1\cr
Date: \tab 2012-04-13\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}

\author{
Daniel-Corneliu Leucuta

Maintainer: Daniel-Corneliu Leucuta <danny.ldc@gmail.com>
}

\keyword{ package }

\seealso{
\code{\link[Rcmdr]{Rcmdr}}
}