\name{RcmdrPlugin.EBM-package}
\alias{RcmdrPlugin.EBM-package}
\alias{RcmdrPlugin.EBM}
\docType{package}
\title{
Evidence Based Medicine Rcmdr Plug-In
}
\description{
This package provides an Rcmdr GUI extension plug-in for Evidence Based Medicine calculations. The purpose of this GUI extension was to improve teaching in the use of Evidence Based Medicine calculations. The interface allows for computing diagnostic medical indicators: sensitivity, specificity, diagnostic accuracy, Youden's index, positive predictive value, negative predictive value, positive likelyhood ratio, negative likelyhood ratio, number needed to diagnose; prognosis medical indicators: relative risk, atributable risk, odds ratio; therapeutic medical indicators: absolute risk reduction, number needed to treat, relative risk reduction, relative risk, odds ratio. Also the interface allows computing the post-test probability from a pre-test probability and the likelyhood ratio.
}
\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.EBM\cr
Type: \tab Package\cr
Version: \tab 1.0-2\cr
Date: \tab 2012-04-13\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}

\author{
Daniel-Corneliu Leucuta

Maintainer: Daniel-Corneliu Leucuta <danny.ldc@gmail.com>
}

\keyword{ package }

\seealso{
\code{\link[Rcmdr]{Rcmdr}}
}